+++ packages/driver-data/ios/RNDriveKitDriverData-Swift.h
@@ -12,4 +12,5 @@
 - (void)getTripsOrderByDateDescWithSynchronizationType:(NSString *_Nonnull)synchronizationType transportationModes:(NSArray *_Nonnull)transportationModes resolver:(RCTPromiseResolveBlock _Nonnull )resolve rejecter:(RCTPromiseRejectBlock _Nonnull )reject;
 - (void)getTripWithItinId:(NSString *_Nonnull)itinId resolver:(RCTPromiseResolveBlock _Nonnull)resolve rejecter:(RCTPromiseRejectBlock _Nonnull)reject;
 - (void)getRouteWithItinId:(NSString *_Nonnull)itinId resolver:(RCTPromiseResolveBlock _Nonnull)resolve rejecter:(RCTPromiseRejectBlock _Nonnull)reject;
+- (void)updateDriverPassengerModeWithItinId:(NSString *_Nonnull)itinId mode:(NSString *_Nonnull)mode comment:(NSString * _Nullable)comment resolver:(RCTPromiseResolveBlock _Nonnull )resolve rejecter:(RCTPromiseRejectBlock _Nonnull )reject;
 @end